requirements.txt
# Larrak Dependencies
#
# IMPORTANT: This project requires conda for proper installation:
#   conda create -n larrak python=3.11
#   conda activate larrak
#   conda install -c conda-forge casadi ipopt numpy scipy matplotlib
#
# This file is auto-generated from environment.yml.
# Run: python scripts/sync_requirements.py --generate
#
# If you must use pip, run: pip install -r requirements.txt
# Then verify installation: python scripts/check_environment.py

bandit
casadi>=3.6.0
colorama
hypothesis>=6.82.0
ipopt>=3.14.0
ipython
joblib>=1.0
jupyter
markdown>=3.5.0
matplotlib>=3.7.0
memory_profiler
mypy>=1.5.0
numpy>=1.24.0
pandas>=2.0.0
pip-audit
plotly>=5.15.0
pygments>=2.16.0
pytest>=7.4.0
pytest-cov>=4.1.0
python-dateutil>=2.8.0
pyyaml>=6.0
radon
redis>=5.0.0
rq>=1.15.0
ruff>=0.0.280
scalene
scikit-learn>=1.1.0
scipy>=1.10.0
typeguard
weaviate-client>=4.0.0
xenon
